+++ appinfo/app.php
@@ -29,6 +29,8 @@ namespace OCA\Nextant\AppInfo;
 require_once __DIR__ . '/autoload.php';
 
 $app = new \OCA\Nextant\AppInfo\Application();
+
 // $app->registerNavigation();
 $app->registerFilesHooks();
 $app->registerSettingsAdmin();
+$app->registerSearchProvider();