+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**v0.3.0**
+* Managing Solr Schema from Administration page or using ./occ nextant:check 
+* bugfixes:
+  - result being displayed twice.
+  - escape few queries to Solr
+
 **v0.2.0**
 * New format: opendocument, office, rtf, html, epub, ...
 * Search within shared file,